backend/requirements.txt
# Backend requirements
youtube-transcript-api
httpx
requests
convex
python-dotenv
fastapi
uvicorn[standard]
yt-dlp
pydantic
pydantic-settings>=2.0.0
openai>=1.0.0

# Additional requirements for Hunter LeadGen
beautifulsoup4>=4.9.3
lxml>=4.6.3
aiofiles>=0.8.0
aiohttp>=3.9.0
redis>=4.0.0
psutil>=5.9.0
aiodns>=3.0.0
charset-normalizer>=3.0.0

# Hunter Pipeline Improvements
rapidfuzz>=1.9.1

# TikTok API - using mock implementation for now
# TikTokApi>=6.2.0  # Requires playwright, which is heavy
# playwright>=1.36.0

# RAG and Embeddings
numpy>=1.24.0
tiktoken>=0.5.0
langchain>=0.1.0
pinecone-client>=2.2.0
weaviate-client>=3.24.0
pyarrow>=14.0.0
nltk>=3.8.0


# Procedural Audio & Dependencies
transformers==4.38.0


# Database dependencies
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0

# Audio processing and preparation
openai-whisper>=20231117
ffmpeg-python>=0.2.0
soundfile>=0.12.1


# Audio source separation and speaker diarization
demucs>=4.0.0
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
funasr

# ── runtime orchestration ──
celery[redis]>=5.3.6
librosa>=0.10.2
# NOTE: transformers moved to Procedural Audio section
fastapi
uvicorn[standard]
torchaudio
soundfile
python-multipart

